04.1-Streamlit_Proyecto/data/requirements.txt
requests==2.31.0
pandas==2.2.2
python-dotenv==1.0.0
openpyxl==3.1.2
psycopg2-binary
sqlalchemy
streamlit
matplotlib==3.8.0
scikit-learn
jupyter
alembic==1.12.1
plotly==5.17.0
numpy>=1.26.4